requirements-web.txt
# Core dependencies for KataGo Python tools
numpy==1.26.4; python_version >= "3.9"
numpy==1.24.4; python_version < "3.9"
torch==2.2.0
sgfmill==1.1.1
chardet==5.2.0

# Dependencies for Real-Time API (Step 2 Implementation)
fastapi==0.109.2
uvicorn==0.27.1
websockets==15.0.1; python_version >= "3.9"
websockets==13.1; python_version < "3.9"
pydantic==2.6.1
orjson==3.9.15
PyYAML==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Testing and Development
pytest==8.0.1
pytest-asyncio
httpx
requests==2.32.4
